refactor(OldSupervisor): Update tests to use OldSupervisor store and adjust imports for ConversationsTable components
